requirements.txt
flask
ultralytics
opencv-python-headless
numpy
Pillow
networkx
Shapely
pytesseract
requests
python-dotenv
